v9.0-7

+++ aws/lambda/package/nodemanager_pkg.sh
@@ -1,7 +1,12 @@
 #!/bin/sh
-rm ../node_manager.zip
+[ -e ../node_manager.zip ] && rm ../node_manager.zip
+[ -e ../version.txt ] && rm ../version.txt
+
+echo `date` > ../version.txt
 
 cd ..
-zip node_manager.zip nodemanager.py utils.py
+zip node_manager.zip nodemanager.py utils.py version.txt
+
+rm version.txt
 
 chmod 777 node_manager.zip
